pyproject.toml
[project]
name = "thekid"
version = "0.1.0"
description = "The Kid - Neuro-Symbolic Specialist"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
